# src/main/resources/com/sonicle/webtop/calendar/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Calendar",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Calendar service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    },
    {
      "auth-apikey-bearer": [],
      "auth-apikey-username": []
    }
  ],
  "paths": {
...
    "/eas/folders/{folderId}/messages-stats": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "List all messages for a specific folder",
        "operationId": "getEasMessagesStats",
        "parameters": [
          {
            "name": "cutoffDate",
            "in": "query",
            "description": "Cut-off date (ISO date/time YYYYMMDD'T'HHMMSS'Z')",
            "schema": {
              "type": "string"
            },
            "example": "20150912T000000Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EasSyncEventStat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          }
        },
        "description": "Returns sync informations for the specified Calendar."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages": {
      "post": {
        "tags": [
          "eas"
        ],
        "summary": "Adds a message",
        "operationId": "addEasMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncEventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncEventStat"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Add new Event into specified Calendar."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages/{id}": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "Get a single message",
        "operationId": "getEasMessage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        },
        "description": "Gets specified Event."
      },
      "put": {
        "tags": [
          "eas"
        ],
        "summary": "Updates a message",
        "operationId": "updateEasMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncEventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EasSyncEventStat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates specified Event."
      },
      "delete": {
        "tags": [
          "eas"
        ],
        "summary": "Deletes a message",
        "operationId": "deleteEasMessage",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        },
        "description": "Deletes specified Event."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        },
        {
          "name": "id",
          "in": "path",
          "description": "Message ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "EasSyncEventStat": {
        "type": "object",
        "description": "Bean for carry event's stat fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Event ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          }
        },
        "required": [
          "id",
          "etag"
        ]
      },
      "EasSyncEvent": {
        "type": "object",
        "description": "Bean for carry event's fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Event ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          },
          "start": {
            "type": "string",
            "description": "Start date/time (ISO format)"
          },
          "end": {
            "type": "string",
            "description": "End date/time (ISO format)"
          },
          "tz": {
            "type": "string",
            "description": "Timezone ID"
          },
          "allDay": {
            "type": "boolean",
            "description": "All day flag"
          },
          "organizer": {
            "type": "string",
            "description": "Organizer address (as RFC822)"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "location": {
            "type": "string",
            "description": "Location"
          },
          "prvt": {
            "type": "boolean",
            "description": "Private flag"
          },
          "busy": {
            "type": "boolean",
            "description": "Busy flag"
          },
          "reminder": {
            "type": "integer",
            "description": "Reminder"
          },
          "recRule": {
            "type": "string",
            "description": "Recurrence RULE string"
          },
          "recStart": {
            "type": "string",
            "description": "Recurrence start date (ISO date YYYYMMDD)"
          },
          "exDates": {
            "type": "array",
            "description": "Excluded dates (ISO date YYYYMMDD)",
            "items": {
              "type": "string"
            }
          },
          "attendees": {
            "type": "array",
            "description": "Appointment attendees",
            "items": {
              "$ref": "#/components/schemas/EasSyncEventDataAttendee"
            }
          }
        },
        "required": [
          "id",
          "etag",
          "start",
          "end",
          "tz",
          "allDay",
          "organizer",
          "prvt",
          "busy",
          "attendees"
        ]
      },
      "EasSyncEventUpdate": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EasSyncEventData"
          },
          "exceptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EasSyncEventData"
            }
          }
        },
        "description": "Bean for carry event's updateable fields"
      },
      "EasSyncEventData": {
        "required": [
          "allDay",
          "attendees",
          "busy",
          "end",
          "organizer",
          "prvt",
          "start",
          "tz"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "description": "Start date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "end": {
            "type": "string",
            "description": "End date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "tz": {
            "type": "string",
            "description": "Timezone ID"
          },
          "allDay": {
            "type": "boolean",
            "description": "All day flag"
          },
          "organizer": {
            "type": "string",
            "description": "Organizer address (as RFC822)"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "location": {
            "type": "string",
            "description": "Location"
          },
          "prvt": {
            "type": "boolean",
            "description": "Private flag"
          },
          "busy": {
            "type": "boolean",
            "description": "Busy flag"
          },
          "reminder": {
            "type": "integer",
            "description": "Reminder"
          },
          "recRule": {
            "type": "string",
            "description": "Recurrence RULE string"
          },
          "recStart": {
            "type": "string",
            "description": "Recurrence start date (ISO date YYYYMMDD)"
          },
          "exDates": {
            "type": "array",
            "description": "Excluded dates (ISO date YYYYMMDD)",
            "items": {
              "type": "string"
            }
          },
          "attendees": {
            "type": "array",
            "description": "Appointment attendees",
            "items": {
              "$ref": "#/components/schemas/EasSyncEventDataAttendee"
            }
          }
        },
        "description": "Bean for carry event's updateable fields"
      },
      "EasSyncEventDataAttendee": {
        "required": [
          "address",
          "role",
          "status",
          "type"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Address (as RFC822)"
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "role": {
            "type": "string",
            "description": "Role"
          },
          "status": {
            "type": "string",
            "description": "Response status"
          }
        },
        "description": "Bean for carry attendee's fields"
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "auth-apikey-bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "API-Key Authentication: specifies the secret token (bearer and username are both required)"
      },
      "auth-apikey-username": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Username",
        "description": "API-Key Authentication: specifies target username (bearer and username are both required)"
      }
    },
...
  },
...
}